+++ src/components/ui/PageFooter.vue
@@ -46,8 +46,7 @@ import SocialLinks from "./SocialLinks.vue";
 </script>
 
 <style scoped lang="less">
-@import url("@style/public/flex.less");
-@import url("@style/public/var.less");
+@import url("@style/public-page.less");
 
 footer {
     width: 100%;