+++ src/views/ChangeLogPage.vue
@@ -12,7 +12,7 @@ import changeLogDoc from "@/assets/docs/CHANGELOG.md";
 </script>
 
 <style scoped lang="less">
-@import url("@style/page/index-public.less");
+@import url("@style/public-page.less");
 
 @padding-top-bottom: 1.5rem;
 @padding-left-right: @padding-top-bottom * 2;