+++ src/views/ShopPage.vue
@@ -1,0 +1,11 @@
+<!-- 小店页面 -->
+
+<template>
+    <section>
+        <slideshow />
+    </section>
+</template>
+
+<script setup lang="ts">
+import slideshow from "@/features/shop/slideshow.vue";
+</script>